refactor: extract extension host kernel
Move local extension loading into a runtime-independent host so future capabilities can share ownership, snapshots, disposal, and diagnostics semantics.
